+++ src/babel/__tests__/examples/stylesheet_mixins_spread.js
@@ -3,19 +3,19 @@ import * as zacs from '@nozbe/zacs'
 const styles = zacs.stylesheet({
   root: {
     // _mixin syntax
-    _mixin: { flex: 1, zIndex: -1000 },
+    ...{ flex: 1, zIndex: -1000 },
     // second mixin with override
-    _mixin: { zIndex: -2 },
+    ...{ zIndex: -2 },
     // nested mixins
-    _mixin: { _mixin: { flexDirection: 'row' }, alignItems: 'center' },
+    ...{ ...{ flexDirection: 'row' }, alignItems: 'center' },
 
     // mixins in nested scoped
     native: {
-      _mixin: { backgroundColor: 'native' },
+      ...{ backgroundColor: 'native' },
     },
     web: {
       '& > div': {
-        _mixin: { backgroundColor: 'web' },
+        ...{ backgroundColor: 'web' },
       },
     },
   },