+++ src/babel/__tests__/test.js
@@ -129,6 +129,7 @@ describe('zacs stylesheets', () => {
     'stylesheet_properties',
     'stylesheet_babelReplacement',
     'stylesheet_mixins',
+    'stylesheet_mixins_spread',
     'stylesheet_css',
     'stylesheet_css_order',
     'stylesheet_shorthand_box',
@@ -204,7 +205,7 @@ describe('zacs stylesheets', () => {
     bad('{css: {}}', 'magic css:')
 
     bad('{a: {[name]:0}}', 'simple strings')
-    bad('{a: {foo:0,...styles}}', 'simple strings')
+    bad('{a: {foo:0,...styles}}', 'simple object literals')
     bad('{a: {foo}}', 'simple strings')
     bad('{a: {"foo":0}}', 'web inner styles')
 
@@ -215,7 +216,7 @@ describe('zacs stylesheets', () => {
     bad('{a: {css:`foo${bar}`}}', 'magic css:')
 
     bad('{a: {web:{foo}}}', 'simple strings')
-    bad('{a: {native:{...styles}}}', 'simple strings')
+    bad('{a: {native:{...styles}}}', 'simple object literals')
     bad('{a: {ios:{foo}}}', 'simple strings')
     bad('{a: {android:{foo}}}', 'simple strings')